+++ problem/bin_packing.go
@@ -19,7 +19,7 @@ func NewBinPacking(variant string, n int) *discrete.Problem {
 
 // Bin Packing
 func binPacking(name string) *discrete.Problem {
-	p, cfg := newBinProblem(name)
+	p, cfg := newBinPartitionProblem(name)
 	if p == nil || cfg == nil {
 		return nil
 	}